throttle/requirements/local.txt
-r base.txt
django-debug-toolbar>=1.0
django-extensions>=1.2.5
